Require hydrated organization-image entities
The existing implementation doesn't work with linked entities, but there isn't really an explicit mention of this, nor are we explicitly checking for it. Added docstrings to make this clearer, and also added a check for `image.href` to both `getImageSrcset` and `getDefaultImageLink`. This leaves it up to the consumer to fetch and hydrate the organization-image entity rather than handling it here.

Going to release this as a patch for two reasons:

1. All current usages that would have broken with the new method would have broken with the old method anyway
2. `srcset` (returned by `getImageSrcset`) is an attribute, which is stringified anyway - whether it's "" (old) or "undefined" (new), it will have the same result of a 404